Split personal access token management into Authify.PersonalAccessTokens context
Extract personal access token CRUD and authentication functions from the
Accounts context into a dedicated Authify.PersonalAccessTokens context
module. Update all callers and tests to use the new context.
